refactor(sleep): move sleep function out of Runs and into `~/server/utils`
