+++ api/server/routes/assistants/chat.js
@@ -11,8 +11,8 @@ const {
 } = require('~/server/services/Threads');
 const { runAssistant, createOnTextProgress } = require('~/server/services/AssistantService');
 const { addTitle, initializeClient } = require('~/server/services/Endpoints/assistants');
-const { sendResponse, sendMessage } = require('~/server/utils');
-const { createRun, sleep } = require('~/server/services/Runs');
+const { sendResponse, sendMessage, sleep } = require('~/server/utils');
+const { createRun } = require('~/server/services/Runs');
 const { getConvo } = require('~/models/Conversation');
 const getLogStores = require('~/cache/getLogStores');
 const { logger } = require('~/config');
