+++ api/server/utils/index.js
@@ -5,6 +5,7 @@ const handleText = require('./handleText');
 const cryptoUtils = require('./crypto');
 const citations = require('./citations');
 const sendEmail = require('./sendEmail');
+const queue = require('./queue');
 const files = require('./files');
 const math = require('./math');
 
@@ -17,5 +18,6 @@ module.exports = {
   removePorts,
   sendEmail,
   ...files,
+  ...queue,
   math,
 };
