+++ api/server/utils/queue.js
@@ -53,6 +53,17 @@ function LB_QueueAsyncCall(asyncFunc, args, callback) {
   }
 }
 
+/**
+ * Delays the execution for a specified number of milliseconds.
+ *
+ * @param {number} ms - The number of milliseconds to delay.
+ * @return {Promise<void>} A promise that resolves after the specified delay.
+ */
+function sleep(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
 module.exports = {
+  sleep,
   LB_QueueAsyncCall,
 };
